requirements.txt
polygon-api-client
python-dotenv
autogen-agentchat
loguru
pyautogen
fastapi[standard]